refactor(backend): split invoke and extension capability boundaries (#820)
