+++ backend/app/integrations/a2a_extensions/service.py
@@ -4,23 +4,55 @@ from __future__ import annotations
 
 import asyncio
 import time
-from dataclasses import dataclass
-from typing import Any, Dict, Literal, Optional, cast
+from typing import Any, Dict, Literal, Optional
 
 from app.core.logging import get_logger
 from app.features.personal_agents.runtime import A2ARuntime
-from app.integrations.a2a_extensions.codex_declaration_diagnostics import (
-    diagnose_codex_discovery_fallback,
+from app.integrations.a2a_extensions.capability_snapshot import (
+    CapabilitySnapshotCacheEntry as _CapabilitySnapshotCacheEntry,
+)
+from app.integrations.a2a_extensions.capability_snapshot import (
+    CompatibilityProfileCapabilitySnapshot,
+    DeclaredMethodCapabilitySnapshot,
+    DeclaredMethodCollectionCapabilitySnapshot,
+    DeclaredSingleMethodCapabilitySnapshot,
+    InterruptCallbackCapabilitySnapshot,
+    InterruptRecoveryCapabilitySnapshot,
+    InvokeMetadataCapabilitySnapshot,
+    ModelSelectionCapabilitySnapshot,
+    ProviderDiscoveryCapabilitySnapshot,
+    RequestExecutionOptionsCapabilitySnapshot,
+    ResolvedCapabilitySnapshot,
+    SessionBindingCapabilitySnapshot,
+    SessionQueryCapabilitySnapshot,
+    StreamHintsCapabilitySnapshot,
+    WireContractCapabilitySnapshot,
+)
+from app.integrations.a2a_extensions.capability_snapshot_builder import (
+    build_codex_discovery_snapshot,
+    build_codex_exec_snapshot,
+    build_codex_review_snapshot,
+    build_codex_thread_watch_snapshot,
+    build_codex_threads_snapshot,
+    build_codex_turns_snapshot,
+    build_compatibility_profile_snapshot,
+    build_declared_method_collection_snapshot,
+    build_interrupt_callback_snapshot,
+    build_interrupt_recovery_snapshot,
+    build_invoke_metadata_snapshot,
+    build_model_selection_snapshot,
+    build_provider_discovery_snapshot,
+    build_request_execution_options_snapshot,
+    build_session_binding_snapshot,
+    build_session_query_snapshot,
+    build_stream_hints_snapshot,
+    build_wire_contract_snapshot,
 )
 from app.integrations.a2a_extensions.codex_discovery_service import (
     CodexDiscoveryService,
 )
-from app.integrations.a2a_extensions.compatibility_profile import (
-    resolve_compatibility_profile,
-)
 from app.integrations.a2a_extensions.contract_utils import (
     as_dict,
-    require_str,
     resolve_jsonrpc_interface,
 )
 from app.integrations.a2a_extensions.errors import (
@@ -28,58 +60,34 @@ from app.integrations.a2a_extensions.errors import (
     A2AExtensionNotSupportedError,
     A2AExtensionUpstreamError,
 )
-from app.integrations.a2a_extensions.interrupt_callback import (
-    resolve_interrupt_callback,
-)
 from app.integrations.a2a_extensions.interrupt_extension_service import (
     InterruptExtensionService,
-)
-from app.integrations.a2a_extensions.interrupt_recovery import (
-    resolve_interrupt_recovery,
 )
 from app.integrations.a2a_extensions.interrupt_recovery_service import (
     InterruptRecoveryService,
 )
-from app.integrations.a2a_extensions.invoke_metadata import resolve_invoke_metadata
-from app.integrations.a2a_extensions.model_selection import resolve_model_selection
 from app.integrations.a2a_extensions.opencode_discovery_service import (
     OpencodeDiscoveryService,
 )
-from app.integrations.a2a_extensions.opencode_provider_discovery import (
-    resolve_opencode_provider_discovery,
-)
 from app.integrations.a2a_extensions.service_common import ExtensionCallResult
-from app.integrations.a2a_extensions.session_binding import resolve_session_binding
 from app.integrations.a2a_extensions.session_extension_service import (
     SessionExtensionService,
 )
 from app.integrations.a2a_extensions.session_query_runtime_selection import (
     ResolvedSessionQueryRuntimeCapability,
-    resolve_runtime_session_query,
-)
-from app.integrations.a2a_extensions.shared_contract import (
-    SUPPORTED_SESSION_BINDING_URIS,
-    SUPPORTED_SESSION_QUERY_URIS,
-    normalize_known_extension_uri,
 )
 from app.integrations.a2a_extensions.shared_support import (
     A2AExtensionSupport,
 )
-from app.integrations.a2a_extensions.stream_hints import resolve_stream_hints
 from app.integrations.a2a_extensions.types import (
-    CompatibilityRetentionEntry,
-    ResolvedCompatibilityProfileExtension,
     ResolvedConditionalMethodAvailability,
     ResolvedInterruptCallbackExtension,
     ResolvedInterruptRecoveryExtension,
     ResolvedInvokeMetadataExtension,
-    ResolvedModelSelectionExtension,
     ResolvedProviderDiscoveryExtension,
     ResolvedSessionBindingExtension,
-    ResolvedStreamHintsExtension,
     ResolvedWireContractExtension,
 )
-from app.integrations.a2a_extensions.wire_contract import resolve_wire_contract
 
 logger = get_logger(__name__)
 _CAPABILITY_SNAPSHOT_CACHE_TTL_SECONDS = 300.0
@@ -90,227 +98,15 @@ _CODEX_DISCOVERY_METHODS = {
     "pluginsRead": "codex.discovery.plugins.read",
     "watch": "codex.discovery.watch",
 }
-_CODEX_DISCOVERY_HUB_CONSUMPTION = {
-    "skillsList": True,
-    "appsList": True,
-    "pluginsList": True,
-    "pluginsRead": True,
-    "watch": False,
-}
-_CODEX_EXEC_METHODS = {
-    "start": "codex.exec.start",
-    "write": "codex.exec.write",
-    "resize": "codex.exec.resize",
-    "terminate": "codex.exec.terminate",
-}
-_CODEX_THREADS_METHODS = {
-    "fork": "codex.threads.fork",
-    "archive": "codex.threads.archive",
-    "unarchive": "codex.threads.unarchive",
-    "metadataUpdate": "codex.threads.metadata.update",
-    "watch": "codex.threads.watch",
-}
 _CODEX_TURNS_METHODS = {
     "steer": "codex.turns.steer",
 }
-_CODEX_TURNS_HUB_CONSUMPTION = {
-    "steer": True,
-}
-_CODEX_REVIEW_METHODS = {
-    "start": "codex.review.start",
-    "watch": "codex.review.watch",
-}
-_CODEX_THREAD_WATCH_METHOD = "codex.threads.watch"
 _CODEX_TURN_CONTROL_URI = "urn:codex-a2a:codex-turn-control/v1"
 _CODEX_TURN_CONTROL_BUSINESS_CODE_MAP = {
     -32007: "authorization_forbidden",
     -32012: "turn_not_steerable",
     -32013: "turn_forbidden",
 }
-_CODEX_REQUEST_EXECUTION_METADATA_FIELD = "metadata.codex.execution"
-
-
-@dataclass(frozen=True, slots=True)
-class DeclaredMethodCapabilitySnapshot:
-    declared: bool
-    consumed_by_hub: bool
-    method: str | None = None
-    availability: Literal["always", "enabled", "disabled", "unsupported"] = (
-        "unsupported"
-    )
-    config_key: str | None = None
-    reason: str | None = None
-    retention: str | None = None
-
-
-@dataclass(frozen=True, slots=True)
-class DeclaredMethodCollectionCapabilitySnapshot:
-    declared: bool
-    consumed_by_hub: bool
-    status: Literal[
-        "unsupported",
-        "declared_not_consumed",
-        "partially_consumed",
-        "supported",
-        "unsupported_by_design",
-    ]
-    methods: dict[str, DeclaredMethodCapabilitySnapshot]
-    jsonrpc_url: str | None = None
-    declaration_source: (
-        Literal[
-            "none",
-            "wire_contract",
-            "wire_contract_fallback",
-            "extension_method_hint",
-            "extension_uri_hint",
-        ]
-        | None
-    ) = None
-    declaration_confidence: Literal["none", "fallback", "authoritative"] | None = None
-    negotiation_state: (
-        Literal["supported", "missing", "invalid", "unsupported"] | None
-    ) = None
-    diagnostic_note: str | None = None
-
-
-@dataclass(frozen=True, slots=True)
-class DeclaredSingleMethodCapabilitySnapshot:
-    declared: bool
-    consumed_by_hub: bool
-    status: Literal["unsupported", "unsupported_by_design"]
-    method: str | None = None
-    jsonrpc_url: str | None = None
-
-
-@dataclass(frozen=True, slots=True)
-class SessionQueryCapabilitySnapshot:
-    status: Literal["supported", "unsupported", "invalid"]
-    capability: ResolvedSessionQueryRuntimeCapability | None = None
-    error: str | None = None
-
-    @property
-    def selection_meta(self) -> dict[str, Any]:
-        if self.capability is None:
-            return {}
-        return {
-            "session_query_declared_contract_family": (
-                self.capability.declared_contract_family
-            ),
-            "session_query_normalized_contract_family": (
-                self.capability.normalized_contract_family
-            ),
-            "session_query_selection_mode": self.capability.selection_mode,
-        }
-
-
-@dataclass(frozen=True, slots=True)
-class SessionBindingCapabilitySnapshot:
-    status: Literal["supported", "unsupported", "invalid"]
-    ext: ResolvedSessionBindingExtension | None = None
-    error: str | None = None
-    meta: dict[str, Any] | None = None
-
-
-@dataclass(frozen=True, slots=True)
-class InvokeMetadataCapabilitySnapshot:
-    status: Literal["supported", "unsupported", "invalid"]
-    ext: ResolvedInvokeMetadataExtension | None = None
-    error: str | None = None
-    meta: dict[str, Any] | None = None
-
-
-@dataclass(frozen=True, slots=True)
-class RequestExecutionOptionsCapabilitySnapshot:
-    status: Literal["unsupported", "declared_not_consumed", "invalid"]
-    declared: bool
-    consumed_by_hub: bool
-    metadata_field: str | None = None
-    fields: tuple[str, ...] = ()
-    persists_for_thread: bool | None = None
-    source_extensions: tuple[str, ...] = ()
-    notes: tuple[str, ...] = ()
-    error: str | None = None
-
-
-@dataclass(frozen=True, slots=True)
-class InterruptCallbackCapabilitySnapshot:
-    status: Literal["supported", "unsupported", "invalid"]
-    ext: ResolvedInterruptCallbackExtension | None = None
-    jsonrpc_url: str | None = None
-    error: str | None = None
-
-
-@dataclass(frozen=True, slots=True)
-class InterruptRecoveryCapabilitySnapshot:
-    status: Literal["supported", "unsupported", "invalid"]
-    ext: ResolvedInterruptRecoveryExtension | None = None
-    jsonrpc_url: str | None = None
-    error: str | None = None
-
-
-@dataclass(frozen=True, slots=True)
-class ProviderDiscoveryCapabilitySnapshot:
-    status: Literal["supported", "unsupported", "invalid"]
-    ext: ResolvedProviderDiscoveryExtension | None = None
-    jsonrpc_url: str | None = None
-    error: str | None = None
-
-
-@dataclass(frozen=True, slots=True)
-class ModelSelectionCapabilitySnapshot:
-    status: Literal["supported", "unsupported", "invalid"]
-    ext: ResolvedModelSelectionExtension | None = None
-    error: str | None = None
-    meta: dict[str, Any] | None = None
-
-
-@dataclass(frozen=True, slots=True)
-class StreamHintsCapabilitySnapshot:
-    status: Literal["supported", "unsupported", "invalid"]
-    ext: ResolvedStreamHintsExtension | None = None
-    error: str | None = None
-    meta: dict[str, Any] | None = None
-
-
-@dataclass(frozen=True, slots=True)
-class CompatibilityProfileCapabilitySnapshot:
-    status: Literal["supported", "unsupported", "invalid"]
-    ext: ResolvedCompatibilityProfileExtension | None = None
-    error: str | None = None
-
-
-@dataclass(frozen=True, slots=True)
-class WireContractCapabilitySnapshot:
-    status: Literal["supported", "unsupported", "invalid"]
-    ext: ResolvedWireContractExtension | None = None
-    error: str | None = None
-
-
-@dataclass(frozen=True, slots=True)
-class ResolvedCapabilitySnapshot:
-    session_query: SessionQueryCapabilitySnapshot
-    session_binding: SessionBindingCapabilitySnapshot
-    invoke_metadata: InvokeMetadataCapabilitySnapshot
-    request_execution_options: RequestExecutionOptionsCapabilitySnapshot
-    interrupt_callback: InterruptCallbackCapabilitySnapshot
-    interrupt_recovery: InterruptRecoveryCapabilitySnapshot
-    model_selection: ModelSelectionCapabilitySnapshot
-    provider_discovery: ProviderDiscoveryCapabilitySnapshot
-    stream_hints: StreamHintsCapabilitySnapshot
-    wire_contract: WireContractCapabilitySnapshot
-    compatibility_profile: CompatibilityProfileCapabilitySnapshot
-    codex_discovery: DeclaredMethodCollectionCapabilitySnapshot
-    codex_threads: DeclaredMethodCollectionCapabilitySnapshot
-    codex_turns: DeclaredMethodCollectionCapabilitySnapshot
-    codex_review: DeclaredMethodCollectionCapabilitySnapshot
-    codex_thread_watch: DeclaredSingleMethodCapabilitySnapshot
-    codex_exec: DeclaredMethodCollectionCapabilitySnapshot
-
-
-@dataclass(slots=True)
-class _CapabilitySnapshotCacheEntry:
-    snapshot: ResolvedCapabilitySnapshot
-    expires_at: float
 
 
 class A2AExtensionsService:
@@ -342,97 +138,15 @@ class A2AExtensionsService:
 
     @staticmethod
     def _build_session_query_snapshot(card: Any) -> SessionQueryCapabilitySnapshot:
-        try:
-            capability = resolve_runtime_session_query(card)
-        except A2AExtensionNotSupportedError as exc:
-            return SessionQueryCapabilitySnapshot(
-                status="unsupported",
-                error=str(exc),
-            )
-        except A2AExtensionContractError as exc:
-            return SessionQueryCapabilitySnapshot(
-                status="invalid",
-                error=str(exc),
-            )
-
-        return SessionQueryCapabilitySnapshot(
-            status="supported",
-            capability=capability,
-        )
+        return build_session_query_snapshot(card)
 
     @staticmethod
     def _build_session_binding_snapshot(card: Any) -> SessionBindingCapabilitySnapshot:
-        try:
-            ext = resolve_session_binding(card)
-        except A2AExtensionNotSupportedError as exc:
-            return SessionBindingCapabilitySnapshot(
-                status="unsupported",
-                error=str(exc),
-                meta={
-                    "session_binding_declared": False,
-                    "session_binding_mode": "compat_fallback",
-                    "session_binding_fallback_used": True,
-                },
-            )
-        except A2AExtensionContractError as exc:
-            return SessionBindingCapabilitySnapshot(
-                status="invalid",
-                error=str(exc),
-                meta={
-                    "session_binding_declared": True,
-                    "session_binding_mode": "compat_fallback",
-                    "session_binding_fallback_used": True,
-                    "session_binding_contract_error": str(exc),
-                },
-            )
-
-        return SessionBindingCapabilitySnapshot(
-            status="supported",
-            ext=ext,
-            meta={
-                "session_binding_declared": True,
-                "session_binding_uri": ext.uri,
-                "session_binding_mode": (
-                    "compat_fallback" if ext.legacy_uri_used else "declared_contract"
-                ),
-                "session_binding_fallback_used": ext.legacy_uri_used,
-            },
-        )
+        return build_session_binding_snapshot(card)
 
     @staticmethod
     def _build_invoke_metadata_snapshot(card: Any) -> InvokeMetadataCapabilitySnapshot:
-        try:
-            ext = resolve_invoke_metadata(card)
-        except A2AExtensionNotSupportedError as exc:
-            return InvokeMetadataCapabilitySnapshot(
-                status="unsupported",
-                error=str(exc),
-                meta={
-                    "invoke_metadata_declared": False,
-                    "invoke_metadata_consumed_by_hub": True,
-                },
-            )
-        except A2AExtensionContractError as exc:
-            return InvokeMetadataCapabilitySnapshot(
-                status="invalid",
-                error=str(exc),
-                meta={
-                    "invoke_metadata_declared": True,
-                    "invoke_metadata_consumed_by_hub": True,
-                    "invoke_metadata_contract_error": str(exc),
-                },
-            )
-
-        return InvokeMetadataCapabilitySnapshot(
-            status="supported",
-            ext=ext,
-            meta={
-                "invoke_metadata_declared": True,
-                "invoke_metadata_consumed_by_hub": True,
-                "invoke_metadata_uri": ext.uri,
-                "invoke_metadata_field_count": len(ext.fields),
-            },
-        )
+        return build_invoke_metadata_snapshot(card)
 
     @staticmethod
     def _normalize_optional_string_list(
@@ -440,363 +154,84 @@ class A2AExtensionsService:
         *,
         field: str,
     ) -> tuple[str, ...]:
-        if value is None:
-            return ()
-        if not isinstance(value, list):
-            raise A2AExtensionContractError(
-                f"Extension contract missing/invalid '{field}'"
-            )
-        items: list[str] = []
-        for index, item in enumerate(value):
-            normalized = require_str(item, field=f"{field}[{index}]")
-            if normalized not in items:
-                items.append(normalized)
-        return tuple(items)
+        from app.integrations.a2a_extensions.capability_snapshot_builder import (
+            normalize_optional_string_list,
+        )
+
+        return normalize_optional_string_list(value, field=field)
 
     @classmethod
     def _build_request_execution_options_snapshot(
         cls,
         card: Any,
     ) -> RequestExecutionOptionsCapabilitySnapshot:
-        capabilities = getattr(card, "capabilities", None)
-        extensions = getattr(capabilities, "extensions", None) if capabilities else None
-        if not extensions:
-            return RequestExecutionOptionsCapabilitySnapshot(
-                status="unsupported",
-                declared=False,
-                consumed_by_hub=False,
-            )
-
-        metadata_field: str | None = None
-        persists_for_thread: bool | None = None
-        fields: list[str] = []
-        notes: list[str] = []
-        source_extensions: list[str] = []
-        found = False
-        supported_sources = {
-            *SUPPORTED_SESSION_BINDING_URIS,
-            *SUPPORTED_SESSION_QUERY_URIS,
-        }
-
-        try:
-            for ext in extensions:
-                raw_uri = getattr(ext, "uri", None)
-                if (
-                    not isinstance(raw_uri, str)
-                    or raw_uri.strip() not in supported_sources
-                ):
-                    continue
-                params = as_dict(getattr(ext, "params", None))
-                if "request_execution_options" not in params:
-                    continue
-                found = True
-                contract = as_dict(params.get("request_execution_options"))
-                if not contract:
-                    raise A2AExtensionContractError(
-                        "Extension contract missing/invalid "
-                        "'params.request_execution_options'"
-                    )
-                current_metadata_field = require_str(
-                    contract.get("metadata_field"),
-                    field="params.request_execution_options.metadata_field",
-                )
-                if current_metadata_field != _CODEX_REQUEST_EXECUTION_METADATA_FIELD:
-                    raise A2AExtensionContractError(
-                        "Extension contract missing/invalid "
-                        "'params.request_execution_options.metadata_field'"
-                    )
-                current_fields = cls._normalize_optional_string_list(
-                    contract.get("fields"),
-                    field="params.request_execution_options.fields",
-                )
-                if not current_fields:
-                    raise A2AExtensionContractError(
-                        "Extension contract missing/invalid "
-                        "'params.request_execution_options.fields'"
-                    )
-                raw_persists_for_thread = contract.get("persists_for_thread")
-                if raw_persists_for_thread is not None and not isinstance(
-                    raw_persists_for_thread, bool
-                ):
-                    raise A2AExtensionContractError(
-                        "Extension contract missing/invalid "
-                        "'params.request_execution_options.persists_for_thread'"
-                    )
-                current_notes = cls._normalize_optional_string_list(
-                    contract.get("notes"),
-                    field="params.request_execution_options.notes",
-                )
-                if metadata_field is None:
-                    metadata_field = current_metadata_field
-                elif metadata_field != current_metadata_field:
-                    raise A2AExtensionContractError(
-                        "Extension contract has conflicting "
-                        "'params.request_execution_options.metadata_field'"
-                    )
-                if raw_persists_for_thread is not None:
-                    if persists_for_thread is None:
-                        persists_for_thread = raw_persists_for_thread
-                    elif persists_for_thread != raw_persists_for_thread:
-                        raise A2AExtensionContractError(
-                            "Extension contract has conflicting "
-                            "'params.request_execution_options.persists_for_thread'"
-                        )
-                for item in current_fields:
-                    if item not in fields:
-                        fields.append(item)
-                for item in current_notes:
-                    if item not in notes:
-                        notes.append(item)
-                normalized_uri = (
-                    normalize_known_extension_uri(raw_uri) or raw_uri.strip()
-                )
-                if normalized_uri not in source_extensions:
-                    source_extensions.append(normalized_uri)
-        except A2AExtensionContractError as exc:
-            return RequestExecutionOptionsCapabilitySnapshot(
-                status="invalid",
-                declared=found,
-                consumed_by_hub=False,
-                metadata_field=metadata_field,
-                fields=tuple(fields),
-                persists_for_thread=persists_for_thread,
-                source_extensions=tuple(source_extensions),
-                notes=tuple(notes),
-                error=str(exc),
-            )
-
-        if not found:
-            return RequestExecutionOptionsCapabilitySnapshot(
-                status="unsupported",
-                declared=False,
-                consumed_by_hub=False,
-            )
-
-        return RequestExecutionOptionsCapabilitySnapshot(
-            status="declared_not_consumed",
-            declared=True,
-            consumed_by_hub=False,
-            metadata_field=metadata_field,
-            fields=tuple(fields),
-            persists_for_thread=persists_for_thread,
-            source_extensions=tuple(source_extensions),
-            notes=tuple(notes),
-        )
+        return build_request_execution_options_snapshot(card)
 
     def _build_interrupt_callback_snapshot(
         self, card: Any
     ) -> InterruptCallbackCapabilitySnapshot:
-        try:
-            ext = resolve_interrupt_callback(card)
-        except A2AExtensionNotSupportedError as exc:
-            return InterruptCallbackCapabilitySnapshot(
-                status="unsupported",
-                error=str(exc),
-            )
-        except A2AExtensionContractError as exc:
-            return InterruptCallbackCapabilitySnapshot(
-                status="invalid",
-                error=str(exc),
-            )
-
-        return InterruptCallbackCapabilitySnapshot(
-            status="supported",
-            ext=ext,
-            jsonrpc_url=self._support.ensure_outbound_allowed(
-                ext.jsonrpc.url, purpose="JSON-RPC interface URL"
-            ),
-        )
+        return build_interrupt_callback_snapshot(self._support, card)
 
     def _build_interrupt_recovery_snapshot(
         self, card: Any
     ) -> InterruptRecoveryCapabilitySnapshot:
-        try:
-            ext = resolve_interrupt_recovery(card)
-        except A2AExtensionNotSupportedError as exc:
-            return InterruptRecoveryCapabilitySnapshot(
-                status="unsupported",
-                error=str(exc),
-            )
-        except A2AExtensionContractError as exc:
-            return InterruptRecoveryCapabilitySnapshot(
-                status="invalid",
-                error=str(exc),
-            )
-
-        return InterruptRecoveryCapabilitySnapshot(
-            status="supported",
-            ext=ext,
-            jsonrpc_url=self._support.ensure_outbound_allowed(
-                ext.jsonrpc.url, purpose="JSON-RPC interface URL"
-            ),
-        )
+        return build_interrupt_recovery_snapshot(self._support, card)
 
     def _build_provider_discovery_snapshot(
         self, card: Any
     ) -> ProviderDiscoveryCapabilitySnapshot:
-        try:
-            ext = resolve_opencode_provider_discovery(card)
-        except A2AExtensionNotSupportedError as exc:
-            return ProviderDiscoveryCapabilitySnapshot(
-                status="unsupported",
-                error=str(exc),
-            )
-        except A2AExtensionContractError as exc:
-            return ProviderDiscoveryCapabilitySnapshot(
-                status="invalid",
-                error=str(exc),
-            )
-
-        return ProviderDiscoveryCapabilitySnapshot(
-            status="supported",
-            ext=ext,
-            jsonrpc_url=self._support.ensure_outbound_allowed(
-                ext.jsonrpc.url, purpose="JSON-RPC interface URL"
-            ),
-        )
+        return build_provider_discovery_snapshot(self._support, card)
 
     @staticmethod
     def _build_model_selection_snapshot(card: Any) -> ModelSelectionCapabilitySnapshot:
-        try:
-            ext = resolve_model_selection(card)
-        except A2AExtensionNotSupportedError as exc:
-            return ModelSelectionCapabilitySnapshot(
-                status="unsupported",
-                error=str(exc),
-                meta={
-                    "model_selection_declared": False,
-                    "model_selection_applies_to_main_chat": False,
-                },
-            )
-        except A2AExtensionContractError as exc:
-            return ModelSelectionCapabilitySnapshot(
-                status="invalid",
-                error=str(exc),
-                meta={
-                    "model_selection_declared": True,
-                    "model_selection_applies_to_main_chat": False,
-                    "model_selection_contract_error": str(exc),
-                },
-            )
-
-        applies_to_main_chat = bool(
-            {"message/send", "message/stream"} & set(ext.applies_to_methods)
-        )
-        return ModelSelectionCapabilitySnapshot(
-            status="supported",
-            ext=ext,
-            meta={
-                "model_selection_declared": True,
-                "model_selection_applies_to_main_chat": applies_to_main_chat,
-                "model_selection_metadata_field": ext.metadata_field,
-            },
-        )
+        return build_model_selection_snapshot(card)
 
     @staticmethod
     def _build_stream_hints_snapshot(card: Any) -> StreamHintsCapabilitySnapshot:
-        try:
-            ext = resolve_stream_hints(card)
-        except A2AExtensionNotSupportedError as exc:
-            return StreamHintsCapabilitySnapshot(
-                status="unsupported",
-                error=str(exc),
-                meta={
-                    "stream_hints_declared": False,
-                    "stream_hints_mode": "compat_fallback",
-                    "stream_hints_fallback_used": True,
-                },
-            )
-        except A2AExtensionContractError as exc:
-            return StreamHintsCapabilitySnapshot(
-                status="invalid",
-                error=str(exc),
-                meta={
-                    "stream_hints_declared": True,
-                    "stream_hints_mode": "compat_fallback",
-                    "stream_hints_fallback_used": True,
-                    "stream_hints_contract_error": str(exc),
-                },
-            )
-
-        return StreamHintsCapabilitySnapshot(
-            status="supported",
-            ext=ext,
-            meta={
-                "stream_hints_declared": True,
-                "stream_hints_uri": ext.uri,
-                "stream_hints_mode": "declared_contract",
-                "stream_hints_fallback_used": False,
-            },
-        )
+        return build_stream_hints_snapshot(card)
 
     @staticmethod
     def _build_compatibility_profile_snapshot(
         card: Any,
     ) -> CompatibilityProfileCapabilitySnapshot:
-        try:
-            ext = resolve_compatibility_profile(card)
-        except A2AExtensionNotSupportedError as exc:
-            return CompatibilityProfileCapabilitySnapshot(
-                status="unsupported",
-                error=str(exc),
-            )
-        except A2AExtensionContractError as exc:
-            return CompatibilityProfileCapabilitySnapshot(
-                status="invalid",
-                error=str(exc),
-            )
-
-        return CompatibilityProfileCapabilitySnapshot(
-            status="supported",
-            ext=ext,
-        )
+        return build_compatibility_profile_snapshot(card)
 
     @staticmethod
     def _build_wire_contract_snapshot(
         card: Any,
     ) -> WireContractCapabilitySnapshot:
-        try:
-            ext = resolve_wire_contract(card)
-        except A2AExtensionNotSupportedError as exc:
-            return WireContractCapabilitySnapshot(
-                status="unsupported",
-                error=str(exc),
-            )
-        except A2AExtensionContractError as exc:
-            return WireContractCapabilitySnapshot(
-                status="invalid",
-                error=str(exc),
-            )
-
-        return WireContractCapabilitySnapshot(
-            status="supported",
-            ext=ext,
-        )
+        return build_wire_contract_snapshot(card)
 
     @staticmethod
     def _declared_wire_contract_methods(
         snapshot: WireContractCapabilitySnapshot,
     ) -> frozenset[str]:
-        if snapshot.status != "supported" or snapshot.ext is None:
-            return frozenset()
-        return frozenset(snapshot.ext.all_jsonrpc_methods)
+        from app.integrations.a2a_extensions.capability_snapshot_builder import (
+            declared_wire_contract_methods,
+        )
+
+        return declared_wire_contract_methods(snapshot)
 
     @classmethod
     def _conditional_wire_contract_methods(
         cls,
         snapshot: WireContractCapabilitySnapshot,
     ) -> dict[str, ResolvedConditionalMethodAvailability]:
-        if snapshot.status != "supported" or snapshot.ext is None:
-            return {}
-        return dict(snapshot.ext.conditionally_available_methods)
+        from app.integrations.a2a_extensions.capability_snapshot_builder import (
+            conditional_wire_contract_methods,
+        )
+
+        return conditional_wire_contract_methods(snapshot)
 
     @staticmethod
     def _compatibility_method_retention(
         snapshot: CompatibilityProfileCapabilitySnapshot,
-    ) -> dict[str, CompatibilityRetentionEntry]:
-        if snapshot.status != "supported" or snapshot.ext is None:
-            return {}
-        return dict(snapshot.ext.method_retention)
+    ) -> dict[str, Any]:
+        from app.integrations.a2a_extensions.capability_snapshot_builder import (
+            compatibility_method_retention,
+        )
+
+        return compatibility_method_retention(snapshot)
 
     @classmethod
     def _resolve_declared_method_snapshot(
@@ -807,52 +242,15 @@ class A2AExtensionsService:
         method_name: str,
         consumed_by_hub: bool,
     ) -> DeclaredMethodCapabilitySnapshot:
-        declared_methods = cls._declared_wire_contract_methods(wire_contract)
-        conditional_methods = cls._conditional_wire_contract_methods(wire_contract)
-        retention_map = cls._compatibility_method_retention(compatibility_profile)
-        conditional = conditional_methods.get(method_name)
-        retention = retention_map.get(method_name)
-
-        active_declared = method_name in declared_methods
-        deployment_conditional = conditional is not None or (
-            retention is not None
-            and retention.retention == "deployment-conditional"
-            and retention.availability in {"enabled", "disabled"}
+        from app.integrations.a2a_extensions.capability_snapshot_builder import (
+            resolve_declared_method_snapshot,
         )
-        declared = active_declared or deployment_conditional
 
-        if active_declared:
-            availability: Literal["always", "enabled", "disabled", "unsupported"] = (
-                cast(
-                    Literal["always", "enabled", "disabled"],
-                    retention.availability,
-                )
-                if retention is not None
-                and retention.availability in {"always", "enabled", "disabled"}
-                else "always"
-            )
-        elif deployment_conditional:
-            availability = (
-                cast(Literal["enabled", "disabled"], retention.availability)
-                if retention is not None
-                and retention.availability in {"enabled", "disabled"}
-                else "disabled"
-            )
-        else:
-            availability = "unsupported"
-
-        return DeclaredMethodCapabilitySnapshot(
-            declared=declared,
-            consumed_by_hub=consumed_by_hub and active_declared,
-            method=method_name if declared else None,
-            availability=availability,
-            config_key=(
-                conditional.toggle
-                if conditional is not None and conditional.toggle
-                else retention.toggle if retention is not None else None
-            ),
-            reason=conditional.reason if conditional is not None else None,
-            retention=retention.retention if retention is not None else None,
+        return resolve_declared_method_snapshot(
+            wire_contract=wire_contract,
+            compatibility_profile=compatibility_profile,
+            method_name=method_name,
+            consumed_by_hub=consumed_by_hub,
         )
 
     @classmethod
@@ -885,41 +283,12 @@ class A2AExtensionsService:
         ) = None,
         diagnostic_note: str | None = None,
     ) -> DeclaredMethodCollectionCapabilitySnapshot:
-        methods = {
-            key: cls._resolve_declared_method_snapshot(
-                wire_contract=wire_contract,
-                compatibility_profile=compatibility_profile,
-                method_name=method_name,
-                consumed_by_hub=bool(hub_consumption.get(key, False)),
-            )
-            for key, method_name in method_map.items()
-        }
-        declared = any(item.declared for item in methods.values())
-        consumed = any(
-            item.declared and item.consumed_by_hub for item in methods.values()
-        )
-        unconsumed = any(
-            item.declared and not item.consumed_by_hub for item in methods.values()
-        )
-        if not declared:
-            status: Literal[
-                "unsupported",
-                "declared_not_consumed",
-                "partially_consumed",
-                "supported",
-                "unsupported_by_design",
-            ] = "unsupported"
-        elif consumed and unconsumed:
-            status = "partially_consumed"
-        elif consumed:
-            status = "supported"
-        else:
-            status = unsupported_status_when_declared
-        return DeclaredMethodCollectionCapabilitySnapshot(
-            declared=declared,
-            consumed_by_hub=consumed,
-            status=status,
-            methods=methods,
+        return build_declared_method_collection_snapshot(
+            wire_contract=wire_contract,
+            compatibility_profile=compatibility_profile,
+            method_map=method_map,
+            hub_consumption=hub_consumption,
+            unsupported_status_when_declared=unsupported_status_when_declared,
             jsonrpc_url=jsonrpc_url,
             declaration_source=declaration_source,
             declaration_confidence=declaration_confidence,
@@ -936,62 +305,11 @@ class A2AExtensionsService:
         *,
         jsonrpc_url: str | None,
     ) -> DeclaredMethodCollectionCapabilitySnapshot:
-        if wire_contract.status == "supported":
-            return cls._build_declared_method_collection_snapshot(
-                wire_contract=wire_contract,
-                compatibility_profile=compatibility_profile,
-                method_map=_CODEX_DISCOVERY_METHODS,
-                hub_consumption=_CODEX_DISCOVERY_HUB_CONSUMPTION,
-                unsupported_status_when_declared="declared_not_consumed",
-                jsonrpc_url=jsonrpc_url,
-                declaration_source="wire_contract",
-                declaration_confidence="authoritative",
-                negotiation_state="supported",
-            )
-
-        fallback = diagnose_codex_discovery_fallback(
+        return build_codex_discovery_snapshot(
             card,
-            wire_contract_status=wire_contract.status,
-        )
-        if fallback.declared:
-            methods = {
-                key: DeclaredMethodCapabilitySnapshot(
-                    declared=method_name in fallback.method_names,
-                    consumed_by_hub=False,
-                    method=(
-                        method_name if method_name in fallback.method_names else None
-                    ),
-                    availability=(
-                        "always"
-                        if method_name in fallback.method_names
-                        else "unsupported"
-                    ),
-                )
-                for key, method_name in _CODEX_DISCOVERY_METHODS.items()
-            }
-            return DeclaredMethodCollectionCapabilitySnapshot(
-                declared=True,
-                consumed_by_hub=False,
-                status="unsupported",
-                methods=methods,
-                jsonrpc_url=None,
-                declaration_source=fallback.source,
-                declaration_confidence=fallback.confidence,
-                negotiation_state=fallback.negotiation_state,
-                diagnostic_note=fallback.note,
-            )
-
-        return cls._build_declared_method_collection_snapshot(
-            wire_contract=wire_contract,
-            compatibility_profile=compatibility_profile,
-            method_map=_CODEX_DISCOVERY_METHODS,
-            hub_consumption=_CODEX_DISCOVERY_HUB_CONSUMPTION,
-            unsupported_status_when_declared="declared_not_consumed",
-            jsonrpc_url=None,
-            declaration_source=fallback.source,
-            declaration_confidence=fallback.confidence,
-            negotiation_state=fallback.negotiation_state,
-            diagnostic_note=fallback.note,
+            wire_contract,
+            compatibility_profile,
+            jsonrpc_url=jsonrpc_url,
         )
 
     @classmethod
@@ -1002,12 +320,9 @@ class A2AExtensionsService:
         *,
         jsonrpc_url: str | None,
     ) -> DeclaredMethodCollectionCapabilitySnapshot:
-        return cls._build_declared_method_collection_snapshot(
-            wire_contract=wire_contract,
-            compatibility_profile=compatibility_profile,
-            method_map=_CODEX_EXEC_METHODS,
-            hub_consumption={},
-            unsupported_status_when_declared="unsupported_by_design",
+        return build_codex_exec_snapshot(
+            wire_contract,
+            compatibility_profile,
             jsonrpc_url=jsonrpc_url,
         )
 
@@ -1019,12 +334,9 @@ class A2AExtensionsService:
         *,
         jsonrpc_url: str | None,
     ) -> DeclaredMethodCollectionCapabilitySnapshot:
-        return cls._build_declared_method_collection_snapshot(
-            wire_contract=wire_contract,
-            compatibility_profile=compatibility_profile,
-            method_map=_CODEX_THREADS_METHODS,
-            hub_consumption={},
-            unsupported_status_when_declared="unsupported_by_design",
+        return build_codex_threads_snapshot(
+            wire_contract,
+            compatibility_profile,
             jsonrpc_url=jsonrpc_url,
         )
 
@@ -1036,12 +348,9 @@ class A2AExtensionsService:
         *,
         jsonrpc_url: str | None,
     ) -> DeclaredMethodCollectionCapabilitySnapshot:
-        return cls._build_declared_method_collection_snapshot(
-            wire_contract=wire_contract,
-            compatibility_profile=compatibility_profile,
-            method_map=_CODEX_TURNS_METHODS,
-            hub_consumption=_CODEX_TURNS_HUB_CONSUMPTION,
-            unsupported_status_when_declared="declared_not_consumed",
+        return build_codex_turns_snapshot(
+            wire_contract,
+            compatibility_profile,
             jsonrpc_url=jsonrpc_url,
         )
 
@@ -1053,12 +362,9 @@ class A2AExtensionsService:
         *,
         jsonrpc_url: str | None,
     ) -> DeclaredMethodCollectionCapabilitySnapshot:
-        return cls._build_declared_method_collection_snapshot(
-            wire_contract=wire_contract,
-            compatibility_profile=compatibility_profile,
-            method_map=_CODEX_REVIEW_METHODS,
-            hub_consumption={},
-            unsupported_status_when_declared="unsupported_by_design",
+        return build_codex_review_snapshot(
+            wire_contract,
+            compatibility_profile,
             jsonrpc_url=jsonrpc_url,
         )
 
@@ -1069,13 +375,8 @@ class A2AExtensionsService:
         *,
         jsonrpc_url: str | None,
     ) -> DeclaredSingleMethodCapabilitySnapshot:
-        declared_methods = cls._declared_wire_contract_methods(wire_contract)
-        declared = _CODEX_THREAD_WATCH_METHOD in declared_methods
-        return DeclaredSingleMethodCapabilitySnapshot(
-            declared=declared,
-            consumed_by_hub=False,
-            status="unsupported_by_design" if declared else "unsupported",
-            method=_CODEX_THREAD_WATCH_METHOD if declared else None,
+        return build_codex_thread_watch_snapshot(
+            wire_contract,
             jsonrpc_url=jsonrpc_url,
         )
